moved frequency domain traits into their own file
