+++ test/main.cpp
@@ -1,3 +1,9 @@
+//
+//  ╔═╗┬ ┬┌┬┐┬┌─┐╔╦╗┬─┐┌─┐┬┌┬┐┌─┐
+//  ╠═╣│ │ ││││ │ ║ ├┬┘├─┤│ │ └─┐
+//  ╩ ╩└─┘─┴┘┴└─┘ ╩ ┴└─┴ ┴┴ ┴ └─┘
+//
+//  © 2021 Lorenz Bucher - all rights reserved
 
 // Let Catch provide main():
 #define CATCH_CONFIG_MAIN
